Update seq2seq_lstm.py
In the previous file, the dataset file was handled as there is only one tab space per line but in reality, the dataset provided by http://www.manythings.org/anki/  has two tab spaces per line that has to be removed. There is a common term '\tCC' in each line that has to be removed to run the dataset. I have added few lines of code to RECTIFY it. You can check at from line 117 to line 119. It worked for me. At least on non-shuffled dataset.